pyproject.toml
[project]
name = "finance-app"
version = "0.1.0"
description = "Personal finance dashboard with Streamlit"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyjwt",
    "cryptography",
    "requests",
    "streamlit",
    "pandas",
    "plotly",
    "anthropic",
    "yfinance",
    "pytr[waf]",
    "pre-commit>=4.5.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/finapp"]

[tool.uv.sources]
pytr = { git = "https://github.com/Felixoid/pytr.git", branch = "fix-aws-waf" }
